requirements.txt
asn1crypto==1.5.1
beautifulsoup4==4.12.3
boto3==1.34.145
botocore==1.34.145
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
jmespath==1.0.1
kafka-python==2.0.2
lxml==5.2.2
numpy==2.0.0
packaging==24.1
pandas==2.2.2
psycopg2==2.9.9
pyodbc==5.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
redshift-connector==2.1.2
requests==2.32.3
s3transfer==0.10.2
scramp==1.4.5
six==1.16.0
soupsieve==2.5
tzdata==2024.1
urllib3==1.26.19
